+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* New command `projectile-run-command-in-root`.
 * New defcustom `projectile-use-git-grep` will make `projectile-grep` use `git grep`
 for git projects.
 * Added new `projectile-commander` methods ?v and ?R which run